move getOrdersByUser to user controller, move getOrdersByShop to shop controller
