+++ backend/src/routes/shop.ts
@@ -3,6 +3,7 @@ import express from 'express';
 import {
     createShop,
     deleteShop,
+    getOrdersByShop,
     getShop,
     getShops,
     updateShop,
@@ -20,5 +21,7 @@ router.post('/', createShop);
 router.put('/:id', updateShop);
 // DELETE /api/shop/:id
 router.delete('/:id', deleteShop);
+// GET /api/shop/:shop_id/orders
+router.get('/:shop_id/orders', getOrdersByShop);
 
 export default router;
